+++ requests/core.py
@@ -12,15 +12,18 @@ This module implements the main Requests system.
 """
 
 __title__ = 'requests'
-__version__ = '0.5.1'
-__build__ = 0x000501
+__version__ = '0.6.1'
+__build__ = 0x000601
 __author__ = 'Kenneth Reitz'
 __license__ = 'ISC'
 __copyright__ = 'Copyright 2011 Kenneth Reitz'
 
 
-from models import HTTPError, auth_manager
+from models import HTTPError, Request, Response
 from api import *
 from exceptions import *
+from sessions import session
 from status_codes import codes
 from config import settings
+
+import utils